+++ CHANGELOG.md
@@ -5,6 +5,9 @@ This document provides a log of changes. It is currently maintained by hand.
 ## (master)
 
 - Fixed index.d.ts declarations missing from the published package...
+- Introduced the `quickstrap` function
+- Improved alias support, fixed some cases where aliases would be ignored
+- Added notes on testing and mocking modules
 - [prefab][mongoose-module] Introduced a prototype prefab for mongoose with typings
 
 ## v0.2